+++ resources/banners.py
@@ -14,16 +14,7 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 """
 
-
-class bcolors(object):
-    HEADER = '\033[95m'
-    OKBLUE = '\033[34m'
-    OKGREEN = '\033[32m'
-    WARNING = '\033[93m'
-    FAIL = '\033[31m'
-    ENDC = '\033[0m'
-    BOLD = '\033[1m'
-    TITLE = '\033[96m'
+from console_colors import bcolors
 
 
 def banner():
