+++ resources/console_colors.py
@@ -1,0 +1,10 @@
+# coding=utf-8
+class bcolors(object):
+    HEADER = '\033[95m'
+    OKBLUE = '\033[34m'
+    OKGREEN = '\033[32m'
+    WARNING = '\033[93m'
+    FAIL = '\033[31m'
+    ENDC = '\033[0m'
+    BOLD = '\033[1m'
+    TITLE = '\033[96m'
